# Move event handlers to `ExpressionContext`