--- a/src/NCalc.Async/AsyncExpressionContext.cs
+++ b/src/NCalc.Async/AsyncExpressionContext.cs
@@ -1,3 +1,5 @@
+using NCalc.Handlers;
+
 namespace NCalc;
 
 public record AsyncExpressionContext : ExpressionContextBase
@@ -8,6 +10,10 @@ public record AsyncExpressionContext : ExpressionContextBase
     public IDictionary<string, AsyncExpressionFunction> Functions { get; set; } =
         new Dictionary<string, AsyncExpressionFunction>();
 
+    
+    public AsyncEvaluateParameterHandler? AsyncEvaluateParameterHandler { get; set; }
+    public AsyncEvaluateFunctionHandler? AsyncEvaluateFunctionHandler { get; set; }
+    
     public AsyncExpressionContext()
     {
     }