--- a/src/NCalc.Async/Services/AsyncEvaluationService.cs
+++ b/src/NCalc.Async/Services/AsyncEvaluationService.cs
@@ -1,5 +1,4 @@
 using NCalc.Domain;
-using NCalc.Handlers;
 using NCalc.Visitors;
 
 namespace NCalc.Services;
@@ -7,14 +6,10 @@ namespace NCalc.Services;
 /// <inheritdoc cref="IAsyncEvaluationService"/>
 public class AsyncEvaluationService : IAsyncEvaluationService
 {
-    public event AsyncEvaluateFunctionHandler? EvaluateFunctionAsync;
-    public event AsyncEvaluateParameterHandler? EvaluateParameterAsync;
-
+    
     public Task<object?> EvaluateAsync(LogicalExpression expression, AsyncExpressionContext context)
     {
         var visitor = new AsyncEvaluationVisitor(context);
-        visitor.EvaluateFunctionAsync += EvaluateFunctionAsync;
-        visitor.EvaluateParameterAsync += EvaluateParameterAsync;
         return expression.Accept(visitor);
     }
 }