--- a/src/NCalc.Async/Visitors/AsyncEvaluationVisitor.cs
+++ b/src/NCalc.Async/Visitors/AsyncEvaluationVisitor.cs
@@ -16,9 +16,6 @@ namespace NCalc.Visitors;
 /// <param name="context">Contextual parameters of the <see cref="LogicalExpression"/>, like custom functions and parameters.</param>
 public class AsyncEvaluationVisitor(AsyncExpressionContext context) : ILogicalExpressionVisitor<Task<object?>>
 {
-    public event AsyncEvaluateFunctionHandler? EvaluateFunctionAsync;
-    public event AsyncEvaluateParameterHandler? EvaluateParameterAsync;
-
     public async Task<object?> Visit(TernaryExpression expression)
     {
         // Evaluates the left expression and saves the value
@@ -166,8 +163,6 @@ public class AsyncEvaluationVisitor(AsyncExpressionContext context) : ILogicalEx
         for (var i = 0; i < argsCount; i++)
         {
             args[i] = new AsyncExpression(function.Expressions[i], context);
-            args[i].EvaluateParameterAsync += EvaluateParameterAsync;
-            args[i].EvaluateFunctionAsync += EvaluateFunctionAsync;
         }
 
         var functionName = function.Identifier.Name;
@@ -212,8 +207,8 @@ public class AsyncEvaluationVisitor(AsyncExpressionContext context) : ILogicalEx
                 foreach (var p in context.DynamicParameters)
                     expression.DynamicParameters[p.Key] = p.Value;
 
-                expression.EvaluateFunctionAsync += EvaluateFunctionAsync;
-                expression.EvaluateParameterAsync += EvaluateParameterAsync;
+                expression.EvaluateFunctionAsync += context.AsyncEvaluateFunctionHandler;
+                expression.EvaluateParameterAsync += context.AsyncEvaluateParameterHandler;
 
                 return await expression.EvaluateAsync();
             }
@@ -238,12 +233,12 @@ public class AsyncEvaluationVisitor(AsyncExpressionContext context) : ILogicalEx
 
     protected ValueTask OnEvaluateFunctionAsync(string name, AsyncFunctionArgs args)
     {
-        return EvaluateFunctionAsync?.Invoke(name, args) ?? default;
+        return context.AsyncEvaluateFunctionHandler?.Invoke(name, args) ?? default;
     }
 
     protected ValueTask OnEvaluateParameterAsync(string name, AsyncParameterArgs args)
     {
-        return EvaluateParameterAsync?.Invoke(name, args) ?? default;
+        return context.AsyncEvaluateParameterHandler?.Invoke(name, args) ?? default;
     }
 
     private async ValueTask<object?> EvaluateAsync(LogicalExpression expression)