--- a/src/NCalc.Sync/Visitors/EvaluationVisitor.cs
+++ b/src/NCalc.Sync/Visitors/EvaluationVisitor.cs
@@ -13,9 +13,6 @@ namespace NCalc.Visitors;
 /// <param name="context">Contextual parameters of the <see cref="LogicalExpression"/>, like custom functions and parameters.</param>
 public class EvaluationVisitor(ExpressionContext context) : ILogicalExpressionVisitor<object?>
 {
-    public event EvaluateFunctionHandler? EvaluateFunction;
-    public event EvaluateParameterHandler? EvaluateParameter;
-
     public object? Visit(TernaryExpression expression)
     {
         // Evaluates the left expression and saves the value
@@ -166,8 +163,6 @@ public class EvaluationVisitor(ExpressionContext context) : ILogicalExpressionVi
         for (var i = 0; i < argsCount; i++)
         {
             args[i] = new Expression(function.Expressions[i], context);
-            args[i].EvaluateParameter += EvaluateParameter;
-            args[i].EvaluateFunction += EvaluateFunction;
         }
 
         var functionName = function.Identifier.Name;
@@ -212,9 +207,9 @@ public class EvaluationVisitor(ExpressionContext context) : ILogicalExpressionVi
                 foreach (var p in context.DynamicParameters)
                     expression.DynamicParameters[p.Key] = p.Value;
 
-                expression.EvaluateFunction += EvaluateFunction;
-                expression.EvaluateParameter += EvaluateParameter;
-
+                expression.EvaluateFunction += context.EvaluateFunctionHandler;
+                expression.EvaluateParameter += context.EvaluateParameterHandler;
+                
                 return expression.Evaluate();
             }
 
@@ -235,11 +230,11 @@ public class EvaluationVisitor(ExpressionContext context) : ILogicalExpressionVi
     }
     protected void OnEvaluateFunction(string name, FunctionArgs args)
     {
-        EvaluateFunction?.Invoke(name, args);
+        context.EvaluateFunctionHandler?.Invoke(name, args);
     }
     protected void OnEvaluateParameter(string name, ParameterArgs args)
     {
-        EvaluateParameter?.Invoke(name, args);
+        context.EvaluateParameterHandler?.Invoke(name, args);
     }
 
     private object? Evaluate(LogicalExpression expression)