--- a/test/NCalc.Benchmarks/Program.cs
+++ b/test/NCalc.Benchmarks/Program.cs
@@ -2,4 +2,5 @@
 using NCalc.Benchmarks;
 
 BenchmarkRunner.Run<LogicalExpressionFactoryBenchmark>(null, args);
+BenchmarkRunner.Run<SimpleEvaluationBenchmark>(null, args);
 BenchmarkRunner.Run<EvaluateVsLambdaBenchmark>(null, args);