Trying to solve #277, did not yield any significant performance changes, but the public API looks cleaner.

Before:
| Method           | Mean     | Error   | StdDev  | Rank | Gen0   | Allocated |
|----------------- |---------:|--------:|--------:|-----:|-------:|----------:|
| SimpleEvaluation | 257.0 ns | 2.32 ns | 2.05 ns |    1 | 0.0820 |     688 B |


After:
| Method           | Mean     | Error   | StdDev  | Rank | Gen0   | Allocated |
|----------------- |---------:|--------:|--------:|-----:|-------:|----------:|
| SimpleEvaluation | 246.3 ns | 4.65 ns | 7.38 ns |    1 | 0.0801 |     672 B |